add panic protection; use channels to register new services
Services and sessions can now panic without bringing down the house of
cards. There's even protection for panic-driven hotloops hogging the
scheduler (discovered using my super artificial, but nonetheless
plausible test cases). Also, the services map was previously being
updated without synchronization, so I have added a serviceRegistration
channel and switched the imFinished channel to be a chan string, which
is also used to pass in the name of the service at registration